refactor model in intermediate api mode
